Relax empty path error detection
We have error detection for preventing deleting the empty path i.e.
'sr d ""'. The empty string is canonicalized as the current working
directory, meaning this will delete the entire cwd, which is almost
certainly a mistake. Thus we detect this and throw an error.

However, the error detection is quite strong and throws an error for
_all_ whitespace paths e.g. 'sr d " "'. This is overzealous as
empty paths are valid POSIX, and whitespace paths are correctly
canonicalized as "cwd/ " rather thant the dangerous "cwd".

Hence we now allow whitespace paths and only error on the empty string.