Remove unnecessary backend golden tests
The Index formatting tests were parameterized on the backend,
as we wanted to ensure that each backend was formatted consistently.
This was unnecessary, however, as the index formatting does not take
backend into consideration at all.

Thus it can be removed, and reduces the number of tests / files by a
factor of 3.