Fix osx integration test
Commit a8ea3b971b50ddcb88f435ee2c58533fd7afe528 had the right idea,
but there was a bug. While the new text normalization logic correctly
forbids identically normalized paths, the base generator,
genFileNameSet, did not prevent duplicate paths in its own set.
It was only the generators that built on this, gen2FileNameSets and
gen3FileNameSets, that preventing generating paths distinct from the
original genFileNameSet.

Thus we retool genFileNameSet so that duplicate paths are removed
from the outset. This is done by changing the basic type we generate
to one that includes the normalized path, and defining equality
based on this (thus it is handled by UniqueSeqNE's fromNonEmpty).

This has the happy side-effect of simplifying the types, too.