Add --indices option to perm-delete and restore
New --indices option that allows a user to delete/restore paths
from the trash based on their numeric index. We first print out
the table with these indices. Consequently, we add a new tabular
formatting method for listing a (simple) table with indices.

We also fix a bug where we reported all non-error paths permanently
deleted, even when the user supplied 'n' (hence the path was not
deleted).